+++ src/user/domains/question/components/QuestionTableHeader.tsx
@@ -2,7 +2,7 @@
 
 function QuestionTableHeader() {
   return (
-    <div>
+    <>
       <TableHeaderLabel className="ml-8">ID</TableHeaderLabel>
       <TableHeaderLabel className="ml-6">세션 명</TableHeaderLabel>
       <TableHeaderLabel className="ml-64">제목</TableHeaderLabel>
@@ -10,7 +10,7 @@ function QuestionTableHeader() {
       <TableHeaderLabel className="ml-32">생성</TableHeaderLabel>
       <TableHeaderLabel className="ml-15">답변</TableHeaderLabel>
       <TableHeaderLabel className="ml-14">상태</TableHeaderLabel>
-    </div>
+    </>
   );
 }
